refactor: move Seed domain to the domain.model package
