+++ src/main/java/br/com/aegro/datasync/seed/application/SeedApplicationService.java
@@ -1,7 +1,6 @@
 package br.com.aegro.datasync.seed.application;
 
 import br.com.aegro.datasync.seed.application.model.SeedModel;
-import br.com.aegro.datasync.seed.domain.Seed;
 
 public interface SeedApplicationService {
