+++ src/main/java/br/com/aegro/datasync/seed/application/converter/SeedConverter.java
@@ -1,7 +1,7 @@
 package br.com.aegro.datasync.seed.application.converter;
 
 import br.com.aegro.datasync.seed.application.model.SeedModel;
-import br.com.aegro.datasync.seed.domain.Seed;
+import br.com.aegro.datasync.seed.domain.model.Seed;
 
 public interface SeedConverter {
