+++ src/main/java/br/com/aegro/datasync/seed/domain/SeedRepository.java
@@ -1,6 +1,6 @@
 package br.com.aegro.datasync.seed.domain;
 
-import br.com.aegro.datasync.user.domain.model.User;
+import br.com.aegro.datasync.seed.domain.model.Seed;
 
 import java.util.List;
