+++ src/main/java/br/com/aegro/datasync/seed/persistence/mapper/SeedMapper.java
@@ -1,6 +1,6 @@
 package br.com.aegro.datasync.seed.persistence.mapper;
 
-import br.com.aegro.datasync.seed.domain.Seed;
+import br.com.aegro.datasync.seed.domain.model.Seed;
 import br.com.aegro.datasync.seed.persistence.model.SeedData;
 
 public interface SeedMapper {
