+++ src/test/java/br/com/aegro/datasync/SeedTests.java
@@ -1,6 +1,6 @@
 package br.com.aegro.datasync;
 
-import br.com.aegro.datasync.seed.domain.Seed;
+import br.com.aegro.datasync.seed.domain.model.Seed;
 import br.com.aegro.datasync.user.domain.model.User;
 import br.com.aegro.datasync.seed.domain.SeedException;
 import org.junit.jupiter.api.Assertions;
